[Chore] #15 - 폴더링 변경
